examples/fastapi/pyproject.toml
[project]
name = "basic-components-fastapi-example"
version = "0.1.0"
requires-python = ">=3.10"

packages = ["components", "docs"]
dependencies = [
    "fastapi[standard]>=0.115.4",
    "jinjax[whitenoise]>=0.47",
    "starlette-wtf>=0.4.5",
    "uvicorn>=0.32.0",
    "wtforms>=3.2.1",
    "jinja2>=3.1.3",
    "arel>=0.3.0",
    "websockets>=13.1",
    "watchfiles>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["components"]
